+++ Foundation/Render/Common/XControl/Pixel.cpp
@@ -18,7 +18,7 @@ Pixel::~Pixel()
 sys::Color4B Pixel::getPixel(float x, float y)
 {
 	uchar data[4] = { 0 };
-	glReadPixels(x, y, 1, 1, GL_RGBA, GL_UNSIGNED_BYTE, data);
+	glReadPixels((int)x, (int)y, 1, 1, GL_RGBA, GL_UNSIGNED_BYTE, data);
 
 	return sys::Color4B(data[0], data[1], data[2], data[3]);
 }
